+++ src/tasks/mod.rs
@@ -6,6 +6,7 @@ mod alacritty;
 mod atlantis;
 mod atom;
 mod bash;
+mod bashprofile;
 mod bashrc;
 mod bazel;
 mod brew;
@@ -58,6 +59,9 @@ mod vscodejson;
 mod windows;
 mod yq;
 mod zsh;
+mod zshenv;
+mod zshlogin;
+mod zshprofile;
 mod zshrc;
 
 pub fn env() -> Exports {
@@ -145,11 +149,12 @@ fn sequence() -> Vec<String> {
         yq::task().name,
         alacritty::task().name, // deps: config
         atom::task().name,
-        bashrc::task().name,  // deps: config
-        inputrc::task().name, // deps: config
-        profile::task().name, // deps: config
-        bash::task().name,    // deps: bashrc,inputrc,profile
-        git::task().name,     // deps: nodejs/npm
+        bashprofile::task().name, // deps: config
+        bashrc::task().name,      // deps: config
+        inputrc::task().name,     // deps: config
+        profile::task().name,     // deps: config
+        bash::task().name,        // deps: bashprofile,bashrc,inputrc,profile
+        git::task().name,         // deps: nodejs/npm
         googlecloudsdk::task().name,
         hyper::task().name, // deps: config
         macos::task().name,
@@ -160,8 +165,11 @@ fn sequence() -> Vec<String> {
         vim::task().name,        // deps: config,pip; takes over the terminal
         vscodejson::task().name, // deps: config
         vscode::task().name,     // deps: config,vscodejson
+        zshenv::task().name,     // deps: config
+        zshlogin::task().name,   // deps: config
+        zshprofile::task().name, // deps: config
         zshrc::task().name,      // deps: config
-        zsh::task().name,        // deps: profile,inputrc,zshrc,brewbundle
+        zsh::task().name,        // deps: profile,zshenv,zshlogin,zshprofile,zshrc,brewbundle
         windows::task().name,
     ]
 }
@@ -171,6 +179,7 @@ fn mapping() -> HashMap<String, Task> {
     map.insert(String::from("alacritty"), alacritty::task());
     map.insert(String::from("atlantis"), atlantis::task());
     map.insert(String::from("atom"), atom::task());
+    map.insert(String::from("bashprofile"), bashprofile::task());
     map.insert(String::from("bashrc"), bashrc::task());
     map.insert(String::from("bash"), bash::task());
     map.insert(String::from("bazel"), bazel::task());
@@ -223,6 +232,9 @@ fn mapping() -> HashMap<String, Task> {
     map.insert(String::from("windows"), windows::task());
     map.insert(String::from("yq"), yq::task());
     map.insert(String::from("zsh"), zsh::task());
+    map.insert(String::from("zshenv"), zshenv::task());
+    map.insert(String::from("zshlogin"), zshlogin::task());
+    map.insert(String::from("zshprofile"), zshprofile::task());
     map.insert(String::from("zshrc"), zshrc::task());
     map
 }